src/event_stream/codec: relocate inline tests to codec_tests.rs
Same playbook. 222-LOC inline test block extracted.
After: codec.rs drops from 581 → 362 LOC.
cargo build/test 92/878.
